+++ qiskit/qobj/qobj.py
@@ -7,16 +7,17 @@
 
 """Model for Qobj."""
 
-from marshmallow.validate import OneOf
+from marshmallow.validate import Equal, OneOf
 
+from qiskit.qobj.models.base import QobjExperimentSchema, QobjConfigSchema, QobjHeaderSchema
+from qiskit.qobj.models.pulse import PulseQobjExperimentSchema, PulseQobjConfigSchema
+from qiskit.qobj.models.qasm import QasmQobjExperimentSchema, QasmQobjConfigSchema
 from qiskit.validation.base import BaseModel, BaseSchema, bind_schema
 from qiskit.validation.fields import Nested, String
-
-from .models import QobjConfigSchema, QobjExperimentSchema, QobjHeaderSchema
 from .utils import QobjType
 
-
 QOBJ_VERSION = '1.1.0'
+
 """Current version of the Qobj schema.
 
 Qobj schema versions:
@@ -28,16 +29,37 @@ Qobj schema versions:
 
 class QobjSchema(BaseSchema):
     """Schema for Qobj."""
-
     # Required properties.
     qobj_id = String(required=True)
+    schema_version = String(required=True, missing=QOBJ_VERSION)
+
+    # Required properties depend on Qobj type.
     config = Nested(QobjConfigSchema, required=True)
     experiments = Nested(QobjExperimentSchema, required=True, many=True)
     header = Nested(QobjHeaderSchema, required=True)
-    type = String(required=True,
-                  validate=OneOf(QobjType.QASM.value,
-                                 QobjType.PULSE.value))
-    schema_version = String(required=True, missing=QOBJ_VERSION)
+    type = String(required=True, validate=OneOf(choices=(QobjType.QASM, QobjType.PULSE)))
+
+
+class QasmQobjSchema(QobjSchema):
+    """Schema for QasmQobj."""
+
+    # Required properties.
+    config = Nested(QasmQobjConfigSchema, required=True)
+    experiments = Nested(QasmQobjExperimentSchema, required=True, many=True)
+
+    type = String(required=True, validate=Equal(QobjType.QASM),
+                  missing=QobjType.QASM)
+
+
+class PulseQobjSchema(QobjSchema):
+    """Schema for PulseQobj."""
+
+    # Required properties.
+    config = Nested(PulseQobjConfigSchema, required=True)
+    experiments = Nested(PulseQobjExperimentSchema, required=True, many=True)
+
+    type = String(required=True, validate=Equal(QobjType.PULSE),
+                  missing=QobjType.PULSE)
 
 
 @bind_schema(QobjSchema)
@@ -52,8 +74,7 @@ class Qobj(BaseModel):
         config (QobjConfig): config settings for the Qobj.
         experiments (list[QobjExperiment]): list of experiments.
         header (QobjHeader): headers.
-        type (str): experiment type (QASM/PULSE).
-        schema_version (str): Qobj version.
+        type (str): Qobj type.
     """
     def __init__(self, qobj_id, config, experiments, header, type, **kwargs):
         # pylint: disable=redefined-builtin
@@ -62,7 +83,58 @@ class Qobj(BaseModel):
         self.experiments = experiments
         self.header = header
         self.type = type
-
         self.schema_version = QOBJ_VERSION
 
         super().__init__(**kwargs)
+
+
+@bind_schema(QasmQobjSchema)
+class QasmQobj(Qobj):
+    """Model for QasmQobj inherit from Qobj.
+
+    Please note that this class only describes the required fields. For the
+    full description of the model, please check ``QasmQobjSchema``.
+
+    Attributes:
+        qobj_id (str): Qobj identifier.
+        config (QASMQobjConfig): config settings for the Qobj.
+        experiments (list[QASMQobjExperiment]): list of experiments.
+        header (QobjHeader): headers.
+    """
+    def __init__(self, qobj_id, config, experiments, header, **kwargs):
+
+        # to avoid specifying 'type' here within from_dict()
+        kwargs.pop('type', None)
+
+        super().__init__(qobj_id=qobj_id,
+                         config=config,
+                         experiments=experiments,
+                         header=header,
+                         type=QobjType.QASM.value,
+                         **kwargs)
+
+
+@bind_schema(PulseQobjSchema)
+class PulseQobj(Qobj):
+    """Model for PulseQobj inherit from Qobj.
+
+    Please note that this class only describes the required fields. For the
+    full description of the model, please check ``PulseQobjSchema``.
+
+    Attributes:
+        qobj_id (str): Qobj identifier.
+        config (PulseQobjConfig): config settings for the Qobj.
+        experiments (list[PulseQobjExperiment]): list of experiments.
+        header (QobjHeader): headers.
+    """
+    def __init__(self, qobj_id, config, experiments, header, **kwargs):
+
+        # to avoid specifying 'type' here within from_dict()
+        kwargs.pop('type', None)
+
+        super().__init__(qobj_id=qobj_id,
+                         config=config,
+                         experiments=experiments,
+                         header=header,
+                         type=QobjType.PULSE.value,
+                         **kwargs)