+++ CHANGELOG.rst
@@ -33,7 +33,7 @@ Added
 - Added function for purity of a mixed state in ``qiskit.quantum_information``
   (#1733)
 - Added parameter to the TextProgressBar to allow the output to be sent to a
-  different output stream 
+  different output stream
 - Added a ``__qiskit_version__`` parameter to the qiskit namespace. This will
   contain a dictionary of versions for all installed qiskit elements. (#1885).
 - Added a ``RunConfig`` object for configurations related to running an
@@ -50,6 +50,8 @@ Added
 - ``execute_circuits()`` and ``assemble_circuits()`` allow setting a qobj_header of type
   QobjHeader to add extra information to the qobj (and thus result).
 - Register indexing supports negative indices (#1875)
+- Added ``nodes_on_wire()`` to DAGCircuit which returns an iterator over all the
+  operations on the given wire
 
 Changed
 -------
@@ -89,6 +91,7 @@ Changed
   rest of spec-defined entities. (#1909).
 - The rzz gate is now represented as a line when printed in text (#1957).
 - Text drawer has support for multi-q gates (#1939).
+- Separate ``Qobj`` into ``PulseQobj`` and ``QasmQobj`` (#1969).
 
 Deprecated
 ----------